Added token check in relation routes

+++ eCommerceAPP/src/main/kotlin/com/example/routes/relationRoutes.kt
@@ -13,6 +13,7 @@ import com.example.security.token.TokenService
 import io.ktor.http.*
 import io.ktor.server.application.*
 import io.ktor.server.auth.*
+import io.ktor.server.auth.jwt.*
 import io.ktor.server.request.*
 import io.ktor.server.response.*
 import io.ktor.server.routing.*
@@ -20,6 +21,7 @@ import io.ktor.server.routing.*
 fun Route.relationRoutes() {
 
     val relationDao = RelationImpl()
+    val userDao = UserImpl()
     authenticate {
         route("/relations") {
             // Order route
@@ -29,44 +31,86 @@ fun Route.relationRoutes() {
                     val userID = call.parameters["userID"]
                     if (userID.isNullOrBlank()) return@get call.respondText("[ERROR] No valid ID has been entered.",
                         status = HttpStatusCode.BadRequest)
+                    call.principal<JWTPrincipal>()?.getClaim("userEmail", String::class)?.let { mail ->
+                        userDao.getUserByEmail(mail)?.let { userInfo ->
 
-                    val orderList = relationDao.getOrders(userID.toInt())
+                            if (userInfo.userID != userID.toInt()){
+                                return@get call.respondText("[ERROR] You can't access to this user's orders.",
+                                    status = HttpStatusCode.Unauthorized)
+                            }
 
-                    return@get call.respond(HttpStatusCode.OK, orderList)
+                            val orderList = relationDao.getOrders(userID.toInt())
+
+                            return@get call.respond(HttpStatusCode.OK, orderList)
+
+                        }
+                    }
+
+
                 }
                 // Post new order
                 post{
-                    val newOrder = call.receiveNullable<Order>() ?: kotlin.run {
+                val userID = call.parameters["userID"]
+                if (userID.isNullOrBlank()) return@post call.respondText("[ERROR] No valid ID has been entered.",
+                    status = HttpStatusCode.BadRequest)
+
+                val newOrder = call.receiveNullable<Order>() ?: kotlin.run {
                         call.respond(HttpStatusCode.BadRequest)
                         return@post
                     }
-                    val addOrder = relationDao.addOrder(newOrder)
 
-                    if (addOrder){
-                        return@post call.respondText("[SUCCESS] Order created", status = HttpStatusCode.Created)
-                    } else {
-                        return@post call.respondText("[ERROR] Order couldn't be created", status = HttpStatusCode.BadRequest)
+                    call.principal<JWTPrincipal>()?.getClaim("userEmail", String::class)?.let { mail ->
+                        userDao.getUserByEmail(mail)?.let { userInfo ->
+
+                            if (userInfo.userID != userID.toInt()) {
+                                return@post call.respondText(
+                                    "[ERROR] You can't create orders for another user.",
+                                    status = HttpStatusCode.Unauthorized
+                                )
+                            }
+                            val addOrder = relationDao.addOrder(newOrder)
+
+                            if (addOrder){
+                                return@post call.respondText("[SUCCESS] Order created", status = HttpStatusCode.Created)
+                            } else {
+                                return@post call.respondText("[ERROR] Order couldn't be created", status = HttpStatusCode.BadRequest)
+                            }
+                        }
                     }
                 }
                 // Put order state
                 put("/{orderID?}") {
                     val orderID = call.parameters["orderID"]
-                    if (orderID.isNullOrBlank()) return@put call.respondText("[ERROR] No valid ID has been entered.",
+                    if (orderID.isNullOrBlank()) return@put call.respondText("[ERROR] No valid order ID has been entered.",
+                        status = HttpStatusCode.BadRequest)
+                    val userID = call.parameters["userID"]
+                    if (userID.isNullOrBlank()) return@put call.respondText("[ERROR] No valid user ID has been entered.",
                         status = HttpStatusCode.BadRequest)
 
                     val newState = call.receiveNullable<OrderState>() ?: kotlin.run {
                         call.respond(HttpStatusCode.BadRequest)
                         return@put
                     }
-                    val putOrder = relationDao.putOrder(orderID.toInt(), newState)
-                    if (putOrder){
-                        return@put call.respondText("[SUCCESS] Order updated to $newState", status = HttpStatusCode.OK)
-                    } else {
-                        return@put call.respondText("[ERROR] Order couldn't be updated", status = HttpStatusCode.BadRequest)
+
+                    call.principal<JWTPrincipal>()?.getClaim("userEmail", String::class)?.let { mail ->
+                        userDao.getUserByEmail(mail)?.let { userInfo ->
+
+                            if (userInfo.userID != userID.toInt()) {
+                                return@put call.respondText(
+                                    "[ERROR] You can't update this user's orders.",
+                                    status = HttpStatusCode.Unauthorized
+                                )
+                            }
+                            val putOrder = relationDao.putOrder(orderID.toInt(), newState)
+                            if (putOrder){
+                                return@put call.respondText("[SUCCESS] Order updated to $newState", status = HttpStatusCode.OK)
+                            } else {
+                                return@put call.respondText("[ERROR] Order couldn't be updated", status = HttpStatusCode.BadRequest)
+                            }
+                        }
                     }
                 }
             }
-
 
             route("/cart/{userID?}"){
                 // Get current cart of user
@@ -77,15 +121,27 @@ fun Route.relationRoutes() {
                             "[ERROR] No valid ID has been entered.",
                             status = HttpStatusCode.BadRequest
                         )
-                    val currentCart = relationDao.getCart(userID.toInt())
+                    call.principal<JWTPrincipal>()?.getClaim("userEmail", String::class)?.let { mail ->
+                        userDao.getUserByEmail(mail)?.let { userInfo ->
 
-                    if (currentCart != null){
+                            if (userInfo.userID != userID.toInt()) {
+                                return@get call.respondText(
+                                    "[ERROR] You can't access to this user's cart.",
+                                    status = HttpStatusCode.Unauthorized
+                                )
+                            }
+                            val currentCart = relationDao.getCart(userID.toInt())
 
-                        call.respond(HttpStatusCode.OK, currentCart)
-                        
-                    } else {
-                        call.respondText("[ERROR] This user has no cart..", status = HttpStatusCode.NotFound)
+                            if (currentCart != null){
+
+                                call.respond(HttpStatusCode.OK, currentCart)
+
+                            } else {
+                                call.respondText("[ERROR] This user has no cart..", status = HttpStatusCode.NotFound)
+                            }
+                        }
                     }
+
                 }
 
                 post {
@@ -99,13 +155,26 @@ fun Route.relationRoutes() {
                         call.respond(HttpStatusCode.BadRequest)
                         return@post
                     }
-                    val addCart = relationDao.createCart(newCart)
+                    call.principal<JWTPrincipal>()?.getClaim("userEmail", String::class)?.let { mail ->
+                        userDao.getUserByEmail(mail)?.let { userInfo ->
 
-                    if (addCart){
-                        return@post call.respondText("[SUCCESS] Cart created", status = HttpStatusCode.Created)
-                    } else {
-                        return@post call.respondText("[ERROR] Cart couldn't be created", status = HttpStatusCode.BadRequest)
+                            if (userInfo.userID != userID.toInt()) {
+                                return@post call.respondText(
+                                    "[ERROR] You can't access to this user's cart.",
+                                    status = HttpStatusCode.Unauthorized
+                                )
+                            }
+                            val addCart = relationDao.createCart(newCart)
+
+                            if (addCart){
+                                return@post call.respondText("[SUCCESS] Cart created", status = HttpStatusCode.Created)
+                            } else {
+                                return@post call.respondText("[ERROR] Cart couldn't be created", status = HttpStatusCode.BadRequest)
+                            }
+                        }
                     }
+
+
                 }
                 // Update product list of cart
                 put {
@@ -119,12 +188,23 @@ fun Route.relationRoutes() {
                         call.respond(HttpStatusCode.BadRequest)
                         return@put
                     }
-                    val updateCart = relationDao.updateCart(userID.toInt(), newProductList)
+                    call.principal<JWTPrincipal>()?.getClaim("userEmail", String::class)?.let { mail ->
+                        userDao.getUserByEmail(mail)?.let { userInfo ->
 
-                    if (updateCart){
-                        return@put call.respondText("[SUCCESS] Cart updated", status = HttpStatusCode.OK)
-                    } else {
-                        return@put call.respondText("[ERROR] Cart couldn't be updated", status = HttpStatusCode.BadRequest)
+                            if (userInfo.userID != userID.toInt()) {
+                                return@put call.respondText(
+                                    "[ERROR] You can't access to this user's cart.",
+                                    status = HttpStatusCode.Unauthorized
+                                )
+                            }
+                            val updateCart = relationDao.updateCart(userID.toInt(), newProductList)
+
+                            if (updateCart){
+                                return@put call.respondText("[SUCCESS] Cart updated", status = HttpStatusCode.OK)
+                            } else {
+                                return@put call.respondText("[ERROR] Cart couldn't be updated", status = HttpStatusCode.BadRequest)
+                            }
+                        }
                     }
                 }
 
@@ -135,21 +215,28 @@ fun Route.relationRoutes() {
                             "[ERROR] No valid ID has been entered.",
                             status = HttpStatusCode.BadRequest
                         )
-                    val cartID = relationDao.getCart(userID.toInt())!!.idCart
-                    val deleteCart = relationDao.deleteCart(cartID)
 
-                    if (deleteCart){
-                        return@delete call.respondText("[SUCCESS] Cart deleted", status = HttpStatusCode.OK)
-                    } else {
-                        return@delete call.respondText("[ERROR] Cart couldn't be deleted", status = HttpStatusCode.BadRequest)
+                    call.principal<JWTPrincipal>()?.getClaim("userEmail", String::class)?.let { mail ->
+                        userDao.getUserByEmail(mail)?.let { userInfo ->
+
+                            if (userInfo.userID != userID.toInt()) {
+                                return@delete call.respondText(
+                                    "[ERROR] You can't delete other user's carts.",
+                                    status = HttpStatusCode.Unauthorized
+                                )
+                            }
+                            val cartID = relationDao.getCart(userID.toInt())!!.idCart
+                            val deleteCart = relationDao.deleteCart(cartID)
+
+                            if (deleteCart){
+                                return@delete call.respondText("[SUCCESS] Cart deleted", status = HttpStatusCode.OK)
+                            } else {
+                                return@delete call.respondText("[ERROR] Cart couldn't be deleted", status = HttpStatusCode.BadRequest)
+                            }
+                        }
                     }
                 }
             }
-
         }
     }
-
-
-
-
 }
